requirements.txt
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.0
Pillow>=9.0.0  # For image processing
colorama==0.4.6
confection==0.1.5
conllu==6.0.0
cymem==2.0.11
Cython==3.0.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fuzzywuzzy==0.18.0
greenlet==3.2.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
medspacy==1.3.1
medspacy-quickumls==3.2
medspacy_unqlite==0.9.6
murmurhash==1.0.12
nltk==3.9.1
nmslib-metabrainz==2.1.3
numpy==1.26.4
packaging==25.0
pluggy==1.5.0
preshed==3.0.9
psutil==7.0.0
pybind11==2.13.6
pydantic==2.11.4
pydantic_core==2.33.2
PyFastNER==1.0.10
Pygments==2.19.1
PyMySQL==1.1.1
PyRuSH==1.0.9
pysbd==0.3.4
pysimstring==1.3.0
pytest==8.3.5
python-dotenv==1.1.0
python-Levenshtein==0.27.1
quicksectx==0.4.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
scispacy==0.5.5
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
thinc==8.2.5
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
Unidecode==1.4.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
